+++ packages/engine/src/networking/functions/validateNetworkObjects.ts
@@ -1,17 +1,17 @@
-import { dispatchAction } from '@xrengine/hyperflux'
-
 import { Engine } from '../../ecs/classes/Engine'
-import { Network, NetworkTopics } from '../classes/Network'
-import { WorldNetworkAction } from './WorldNetworkAction'
+import { World } from '../../ecs/classes/World'
+import { Network } from '../classes/Network'
+import { NetworkPeerFunctions } from './NetworkPeerFunctions'
 
-export async function validateNetworkObjects(network: Network): Promise<void> {
+export async function validateNetworkObjects(world: World, network: Network): Promise<void> {
   for (const [userId, client] of network.peers) {
     if (userId === Engine.instance.userId) continue
     // Validate that user has phoned home recently
     if (Date.now() - client.lastSeenTs > 30000) {
       console.log('Removing client ', userId, ' due to inactivity')
 
-      dispatchAction(WorldNetworkAction.destroyPeer({ $from: userId }), NetworkTopics.world)
+      NetworkPeerFunctions.destroyPeer(network, userId, false, world)
+      network.updatePeers()
 
       console.log('Disconnected Client:', client.userId)
       if (client?.instanceRecvTransport) {
